backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
bcrypt==4.2.1
click==8.1.8
fastapi==0.115.6
h11==0.14.0
httptools==0.6.4
idna==3.10
iniconfig==2.0.0
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.4
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1